pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pyzlc"
version = "1.0.2"
description="A cross-environment tool to publish objects from simulation for Augmented Reality and Human Robot Interaction."
dependencies = [
    "msgpack",
    "pyzmq",
    "colorama",
]
authors = [
  { name = "Xinkai Jiang", email = "jiangxinkai98@gmail.com" },
]
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
]
readme = "README.md"

[tool.setuptools.packages.find]
where = ["src"]
include = ["pyzlc"]